run only if data present

+++ seeder/teardown.ts
@@ -2,7 +2,10 @@ import { prisma } from '.'
 
 
 export const clearUsers = async () => {
+    const response = await prisma.user.findMany()
+    if(response.length > 0){
     await prisma.password.deleteMany()
     await prisma.profile.deleteMany()
     await prisma.user.deleteMany()
+    }
 }
